Add test for some android specific code
Tidy tests into the appropriate files
